refactor: Cleanup

+++ tests/Tonga.Tests/Primitives/CharOfTest.cs
@@ -10,33 +10,35 @@ namespace Tonga.Tests.Primitives
         public void CharOfInteger()
         {
             Assert.True(
-                new CharOf((int)1337).Value() == (char)((int)1337));
+                new CharOf(1337).Value() == (char)1337);
         }
 
         [Fact]
         public void ChardOfIntegerOutsideOfRange()
         {
-            Assert.Throws<OverflowException>(() => new CharOf((int)2147483647).Value());
+            Assert.Throws<OverflowException>(() => new CharOf(2147483647).Value());
         }
 
         [Fact]
         public void CharOfUInteger()
         {
             Assert.True(
-                new CharOf((uint)1337).Value() == (char)((uint)1337));
+                new CharOf((uint)1337).Value() == (char)1337);
         }
 
         [Fact]
         public void ChardOfUIntegerOutsideOfRange()
         {
-            Assert.Throws<OverflowException>(() => new CharOf((uint)2147483647).Value());
+            Assert.Throws<OverflowException>(() =>
+                new CharOf((uint)2147483647).Value()
+            );
         }
 
         [Fact]
         public void CharOfLong()
         {
             Assert.True(
-                new CharOf((long)1338).Value() == (char)((long)1338));
+                new CharOf((long)1338).Value() == (char)1338);
         }
 
         [Fact]
@@ -49,7 +51,7 @@ namespace Tonga.Tests.Primitives
         public void CharOfULong()
         {
             Assert.True(
-                new CharOf((ulong)1338).Value() == (char)((ulong)1338));
+                new CharOf((ulong)1338).Value() == (char)1338);
         }
 
         [Fact]
@@ -62,27 +64,27 @@ namespace Tonga.Tests.Primitives
         public void CharOfShort()
         {
             Assert.True(
-                new CharOf((short)13).Value() == (char)((short)13));
+                new CharOf((short)13).Value() == (char)13);
         }
 
         [Fact]
         public void CharOfUShort()
         {
             Assert.True(
-                new CharOf((ushort)1338).Value() == (char)((ushort)1338));
+                new CharOf((ushort)1338).Value() == (char)1338);
         }
 
         [Fact]
         public void CharOfDouble()
         {
             Assert.True(
-                new CharOf((double)1338.0).Value() == (char)Convert.ToInt64((double)1338.0));
+                new CharOf(1338.0).Value() == (char)Convert.ToInt64(1338.0));
         }
 
         [Fact]
         public void ChardOfDoubleOutsideOfRange()
         {
-            Assert.Throws<OverflowException>(() => new CharOf((double)2147483647.0).Value());
+            Assert.Throws<OverflowException>(() => new CharOf(2147483647.0).Value());
         }
 
         [Fact]
@@ -109,14 +111,14 @@ namespace Tonga.Tests.Primitives
         public void CharOfByte()
         {
             Assert.True(
-                new CharOf((byte)13).Value() == (char)((byte)13));
+                new CharOf((byte)13).Value() == (char)13);
         }
 
         [Fact]
         public void CharOfSByte()
         {
             Assert.True(
-                new CharOf((sbyte)13).Value() == (char)((sbyte)13));
+                new CharOf((sbyte)13).Value() == (char)13);
         }
     }
 }
